some comments
