+++ packages/flame/test/raster_sprite_test.dart
@@ -1,5 +1,6 @@
 import 'dart:ui';
 
+import 'package:flame/cache.dart';
 import 'package:flame/components.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter/material.dart';
@@ -8,24 +9,75 @@ import 'package:flutter_test/flutter_test.dart';
 import '_resources/load_image.dart';
 
 void main() {
-  group('Sprite', () {
-    testGolden(
-      'Render with anchor',
-      (game) async {
-        game.add(_MyComponent()..position = Vector2.all(25));
-      },
-      goldenFile: '_goldens/sprite_test_1.png',
-    );
+  group('Raster Sprite', () {
+    group('rasterize sprite', () {
+      testGolden(
+        'still renders correctly',
+        (game) async {
+          game.add(_MyRasterComponent()..position = Vector2.all(25));
+        },
+        goldenFile: '_goldens/sprite_test_1.png',
+      );
+    });
+
+    test('adds the image to the cache', () async {
+      final images = Images();
+      expect(images.keys, isEmpty);
+
+      final image = await loadImage('flame.png');
+      images.add('flame.png', image);
+
+      final baseSprite = await Sprite.load(
+        'flame.png',
+        images: images,
+      );
+      await baseSprite.rasterize(images: images);
+
+      expect(images.keys, isNotEmpty);
+
+      // A second rasterization with the same image should not add a new entry
+      final secondBaseSprite = await Sprite.load(
+        'flame.png',
+        images: images,
+      );
+      await secondBaseSprite.rasterize(images: images);
+      expect(images.keys, hasLength(2));
+    });
+
+    group('when using a custom cache key', () {
+      test('adds the image to the cache with the custom key', () async {
+        final images = Images();
+        expect(images.keys, isEmpty);
+
+        final image = await loadImage('flame.png');
+        images.add('flame.png', image);
+
+        final baseSprite = await Sprite.load(
+          'flame.png',
+          images: images,
+        );
+        const customKey = 'custom_key';
+        await baseSprite.rasterize(cacheKey: customKey, images: images);
+
+        expect(images.keys, contains(customKey));
+
+        // A second rasterization with the same image and custom key should not
+        // add a new entry
+        await baseSprite.rasterize(cacheKey: customKey, images: images);
+        expect(images.keys, hasLength(2));
+      });
+    });
   });
 }
 
-class _MyComponent extends PositionComponent {
-  _MyComponent() : super(size: Vector2(200, 400));
+class _MyRasterComponent extends PositionComponent {
+  _MyRasterComponent() : super(size: Vector2(200, 400));
   late final Sprite sprite;
 
   @override
   Future<void> onLoad() async {
-    sprite = Sprite(await loadImage('flame.png'));
+    final baseSprite = Sprite(await loadImage('flame.png'));
+    sprite = await baseSprite.rasterize();
   }
 
   @override
